+++ policy_sentry/shared/conditions.py
@@ -1,3 +1,4 @@
+# Just some text transformation functions related to our conditions-related workarounds
 def get_service_from_condition_key(condition_key):
     elements = condition_key.split(':', 2)
     return elements[0]
@@ -9,5 +10,4 @@ def get_comma_separated_condition_keys(condition_keys):
     :return: result: String containing multiple condition keys, comma-separated
     """
     result = condition_keys.replace('  ', ',')  # replace the double spaces with a comma
-
     return result