+++ CHANGELOG.md
@@ -1,14 +1,20 @@
 # Changelog
 ## 2019-11-07
 ### Added
-* Query capabilities to address #29. This currently includes capabilites for:
-  - Action Table: (1) list of all actions in a service, (2) details on specific action
+* Query capabilities to address #29. This currently includes capabilities for:
+  - Action Table: 
+    - (1) list of all actions in a service
+    - (2) details on specific action
+    - (3) details on actions matching a service and access levels
     - Need to provide more granular queries for querying the access table before finalizing PR. Especially the one based on access levels, and available condition keys per action or ARN.
   - ARN Table: (1) list of all raw ARNs in a service, (2) details about a specific ARN type
   - Conditions Table: (1) list of all condition keys available to a service, (2) details about a specific condition key
 * Unit tests to accompany the query capabilities
+
 ### Changed
 * Fixed naming of a few unit tests to improve output in nosetests.
+* Fixed Condition Keys in actions table; it was previously set to the string 'None' instead of a null value. 
+* Fixed #33 by adding lines 214-222 to database.py
 
 ## 2019-10-24
 ### Added